requirements.txt
loguru
fastapi
uvicorn
openai
tiktoken
pydantic
httpx
python-multipart
gradio
pyyaml
